Factor out `AutoShaderSwitcher` from `ShaderManager`
