+++ src/gui/CMakeLists.txt
@@ -5,6 +5,7 @@ target_sources(libdosboxcommon PRIVATE
   titlebar.cpp
 
   render/auto_image_adjustments.cpp
+  render/auto_shader_switcher.cpp
   render/deinterlacer.cpp
   render/opengl_renderer.cpp
   render/render.cpp
